requirements.txt
django>=4.2
djangorestframework>=3.14
psycopg2-binary>=2.9
django-cors-headers>=4.3
python-decouple>=3.8